VCL-1103 - change vcldsemaphore.pid from smallint to mediumint
vcl.sql: changed definition of vcldsemaphore.pid to be mediumint instead of smallint

update-vcl.sql:
-changed definition of vcldsemaphore.pid to be mediumint instead of smallint
-added alter table to change vcldsemaphore.pid to be mediumint instead of smallint

+++ mysql/vcl.sql
@@ -1481,7 +1481,7 @@ CREATE TABLE IF NOT EXISTS `variable` (
 CREATE TABLE IF NOT EXISTS `vcldsemaphore` (
   `identifier` varchar(256) NOT NULL,
   `reservationid` mediumint(9) unsigned NOT NULL,
-  `pid` smallint(5) unsigned NOT NULL,
+  `pid` mediumint(8) unsigned NOT NULL,
   `expires` datetime NOT NULL,
   PRIMARY KEY (`identifier`),
   KEY `reservationid` (`reservationid`)
